- added offset to formatter - reworked and tested nand dump process - added retries and checks for failed read/write operations
NOTE:
Don't forget to comment out redirection when testing nand dumping. SLC dump has many read errors with 0xFFFAFFFE code. It could be bad sectors but the amount of those is too big for that in my opinion. So maybe some permissions stuff.

TODO:
Write some marker in a predefined position on SD which should be read on boot and depending on the result either redNAND or dumper should be executed